Factor out simulate functions to `PriceMath`
This will allow the frontend to use the simulate function without
needing to instantiate the `AmmClient`.
